# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.User:
    properties:
      about:
        type: string
      address:
        type: string
      avatar:
        type: string
      birthday:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      login_date:
        type: string
      password:
        type: string
      phone_number:
        type: string
      postcode:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - first_name
    - last_name
    - password
    type: object
  models.UsersList:
    properties:
      has_more:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
info:
  contact:
    email: jaganathan.eswaran@gmail.com
    name: Dhuruvah Apps
    url: https://github.com/dhuruvah-apps/BuildSMEs-Api
  description: Example Golang REST API
  title: Go Example REST API
  version: "1.0"
paths:
  /auth/{id}:
    delete:
      consumes:
      - application/json
      description: some description
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Delete user account
      tags:
      - Auth
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: get user by id
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: update existing user
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Update user
      tags:
      - Auth
  /auth/{id}/avatar:
    post:
      consumes:
      - application/json
      description: Post user avatar image
      parameters:
      - description: Body with image file
        in: formData
        name: file
        required: true
        type: file
      - description: aws s3 bucket
        format: bucket
        in: query
        name: bucket
        required: true
        type: string
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Post avatar
      tags:
      - Auth
  /auth/all:
    get:
      consumes:
      - application/json
      description: Get the list of all users
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get users
      tags:
      - Auth
  /auth/find:
    get:
      consumes:
      - application/json
      description: Find user by name
      parameters:
      - description: username
        format: username
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Find by name
      tags:
      - Auth
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get current user by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get user by id
      tags:
      - Auth
...
  /auth/token:
    get:
      consumes:
      - application/json
      description: Get CSRF token, required auth session cookie
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get CSRF token
      tags:
      - Auth
...
swagger: "2.0"